Refactor example CurveFitAnalysis class to use library functions
